text_output: move to lib/
