+++ lib/text_output.py
